terraform/scripts/cloud_run_api/requirements.txt
flask==3.*
google-cloud-bigquery==3.*
gunicorn==22.*
